refactor(agent): merge file-lock into execution-graph
