+++ packages/agent/src/execution/execution-supervisor.ts
@@ -15,7 +15,7 @@ import {
   reviewTaskResult,
   rotateAgent,
   sendReviewFeedback,
-} from "./review";
+} from "./execution-review";
 import type {
   ChildRunResult,
   DependencyGraph,
