+++ packages/agent/src/task/manager.ts
@@ -1,3 +1,8 @@
+// TaskManager: ~500 LOC — tightly coupled task lifecycle orchestrator.
+// Manages creation, triggering, state transitions, and run coordination
+// as a single cohesive unit. Splitting would thread shared task state
+// through multiple function parameters (see modular-code-architecture Rule 5).
+
 import { Task } from "./types";
 import { TaskStorage, TaskListFilter } from "./storage";
 import { TaskStatusManager } from "./lifecycle";
@@ -7,7 +12,26 @@ export type { TriggerError, TriggerResult } from "./trigger-engine";
 import { Bus } from "@openomni/session";
 import { Task as TaskEvent } from "@openomni/protocol";
 import { randomUUID } from "crypto";
-import { PolicyError } from "./errors";
+
+/**
+ * PolicyError — policy violation errors for task execution hardening
+ *
+ * Used to identify and handle policy violations in task execution:
+ * - D6_task_from_task: trigger_task blocked in task context
+ * - D6_task_creation: TaskManager.create blocked in task context
+ * - anti_loop_self_retrigger: Completion event self-retrigger blocked
+ */
+export class PolicyError extends Error {
+  public readonly name = "PolicyError" as const;
+
+  constructor(
+    public readonly code: string,
+    message: string,
+  ) {
+    super(message);
+    this.name = "PolicyError";
+  }
+}
 
 function isActiveRunStatus(status: Task.Run["status"]): boolean {
   return status === "scheduled" || status === "running" || status === "blocked";
